fix!: certificate `required_weight` to use full committee weights (#266)

+++ crates/hashi/src/mpc/types.rs
@@ -426,7 +426,7 @@ pub struct DealerFlowData {
     pub recipients: Vec<Address>,
     pub messages_hash: DealerMessagesHash,
     pub my_signature: MemberSignature,
-    pub required_weight: u16,
+    pub required_weight: u64,
     pub committee: Committee,
 }
